+++ app/Http/Controllers/CategoryController.php
@@ -7,9 +7,12 @@ use Illuminate\Http\Request;
 
 class CategoryController extends Controller
 {
-    
+
     public function create(){
-        return view('dashboard.category');
+        $categories = Category::all();
+        return view('dashboard.category' , [
+            'categories' => $categories
+        ]);
     }
 
     public function insert(Request $request){
@@ -29,6 +32,48 @@ class CategoryController extends Controller
     }
 
     public function update(Request $request){
+        try {
 
+
+
+                $request->validate([
+
+                    'old_name' => 'string|min:4|exists:categories,name',
+                    'new_name' => "string|min:4|unique:categories,name|different:old_name" ,
+                ]);
+                $category = Category::where('name' , $request->old_name)->first();
+                $category->name = $request->new_name ;
+                $category->save();
+                $request->session()->put('update_category' , trans('Your Category updated'));
+                return redirect()->back();
+
+
+
+
+        } catch (\Throwable $th) {
+           return $th->getMessage();
+        }
+
+    }
+
+    public function delete(Request $request , $id){
+        try {
+            if($id == null){
+                throw new \Exception("ID category can't be null ");
+            }else{
+                // check if id is valid
+                $category  = Category::find($id);
+                if($category){
+                    $name = $category->name;
+                    $category->delete();
+                    $request->session()->put('delete-category' , trans("Category [:name] Deleted" , ['name' => $name]));
+                    return redirect()->back();
+                }else{
+                    throw new \Exception("INVALID CATEGORY ID NUMBER , NOT FOUND");
+                }
+            }
+        } catch (\Throwable $th) {
+            return $th->getMessage();
+        }
     }
 }