+++ app/View/Components/Partials/Category/CategoryItemList.php
@@ -1,0 +1,29 @@
+<?php
+
+namespace App\View\Components\Partials\Category;
+
+use App\Models\Category;
+use Closure;
+use Illuminate\Contracts\View\View;
+use Illuminate\View\Component;
+
+class CategoryItemList extends Component
+{
+    /**
+     * Create a new component instance.
+     */
+    public function __construct(
+        public Category $category,
+    )
+    {
+        //
+    }
+
+    /**
+     * Get the view / contents that represent the component.
+     */
+    public function render(): View|Closure|string
+    {
+        return view('components.partials.category.category-item-list');
+    }
+}