+++ database/seeders/DatabaseSeeder.php
@@ -14,12 +14,14 @@ class DatabaseSeeder extends Seeder
      */
     public function run(): void
     {
-        // User::factory(10)->create();
 
-        User::factory()->create([
-            'name' => 'admin',
-            'email' => 'admin@gmail.com',
-            'password' => Hash::make('password'),
+
+
+
+        $this->call([
+            UserSeeder::class ,
+            CategorySeeder::class,
+
         ]);
     }
 }