+++ resources/views/components/tabs/tab-sections-container.blade.php
@@ -1,0 +1,7 @@
+<div id="default-tab-content">
+    {{
+        $slot
+    }}
+    
+
+</div>