+++ tests/Unit/Category/CreateTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Tests\Unit\Category;
+
+use PHPUnit\Framework\TestCase;
+
+class CreateTest extends TestCase
+{
+    /**
+     * A basic unit test example.
+     */
+    public function create_category(): void
+    {
+        // $response = $this->get('dashboard/category/create');
+        // $response->assertStatus(200);
+    }
+}