+++ changeLog.md
@@ -1,4 +1,4 @@
 # V1
 - Auth with  Starter Kit
-
+- Manage (meaning, create/update/delete) categories